python/requirements.txt
bson==0.5.10
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dataclasses==0.8
dnspython==2.1.0
fastapi==0.63.0
h11==0.12.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
Pillow==8.1.1
pkg-resources==0.0.0
pydantic==1.8.1
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-multipart==0.0.5
requests==2.25.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.3
uuid==1.30
uvicorn==0.13.4
Wave==0.0.2
